Rebase - Conflict Refactor and adjust UI according to visily design
