+++ src/components/Commission/CommissionSchemeCreateModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import api from '../api/axios';
+import api from '../../api/axios';
 import PropTypes from "prop-types";
 
 const CommissionSchemeCreateModal = ({onClose, onCreated}) => {
