+++ src/components/Commission/CommissionSchemeEditModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import api from '../api/axios';
+import api from '../../api/axios';
 import PropTypes from "prop-types";
 
 const CommissionSchemeEditModal = ({ selectedScheme, onClose, onUpdated }) => {
