+++ src/components/Commission/CompanyCommissionSchemeCreateModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import api from '../api/axios';
+import api from '../../api/axios';
 import PropTypes from "prop-types";
